Rename suby module to run function for better API clarity
