+++ suby/__init__.py
@@ -1,9 +1,4 @@
-import sys
-
-from suby.proxy_module import ProxyModule as ProxyModule
+from suby.run import run as run
 from suby.errors import RunningCommandError as RunningCommandError, WrongCommandError as WrongCommandError  # noqa: F401
 
 from cantok import TimeoutCancellationError as TimeoutCancellationError  # noqa: F401
-
-
-sys.modules[__name__].__class__ = ProxyModule
